(1)  Added highlighting of the "sticky day" to the mini-cals, on top of highlighting the current day/week/month.
(2)  Implemented the "sticky day" in the week links, preserving the day of the week when moving to a week for which the current sticky day is invalid.

(3)  Added links to the View Day/Week/Month labels in the classic trailer, to bring it into line with the simple trailer and to assist navigation between the day/week/month views now that the "sticky date" concept is working.


git-svn-id: svn://svn.code.sf.net/p/mrbs/code/mrbs/trunk@918 25f7e783-145e-4f2d-a9f8-dd7182bfe9c7