Upgraded the DB authentication scheme so that access rights are now stored in the MRBS database and administered from MRBS.   (Previously admin users were configured in the config file).
If upgrading from an earlier version of MRBS that was using DB authentication, then the first time that edit_users.php is run, MRBS will automatically add the access rights field to the users table and give admin rights to those users for whom admin rights are configured in the config file.   So if upgrading, make sure that the admin rights are in the config file until edit_users.php has been run for the first time.   After that MRBS does not check the config file for admin rights, but instead checks the database, so the admin rights in the config file can then be deleted.

If running MRBS, or MRBS using DB authentication for the first time, the system will automatically create the users table as before and then you must add the first administrator user.

As safety measures, MRBS will force the first user created to be an admin user, and will prevent you from deleting the last admin user in the table.

At the same time the function authGetUserLevel has been modified so that it just takes one parameter instead of two.   The second parameter was always the same, $auth, which was a global anyway.   With the changes to the DB authentication scheme, the second parameter is not needed, and for all other schemes it can be accessed as a global.

Based on contributions from Taryn East.

git-svn-id: svn://svn.code.sf.net/p/mrbs/code/mrbs/trunk@1029 25f7e783-145e-4f2d-a9f8-dd7182bfe9c7